+++ cw_zcash/lib/src/zcash_wallet.dart
@@ -1160,7 +1160,7 @@ abstract class ZcashWalletBase
     final bal =
         balance[CryptoCurrency.zec]!.confirmed +
         balance[CryptoCurrency.zec]!.unconfirmed +
-        balance[CryptoCurrency.zec]!.frozen;
+        balance[CryptoCurrency.zec]!.frozen.toInt();
     final osCacheDir = await getApplicationCacheDirectory();
     final cacheDir = osCacheDir.createTempSync("zkool-import");
     zkoolSweep = ZkoolSweep(
